fix(updatebot):  custom remote branch name
Resolves remote branch name at runtime using a combination of .updatebot.yml setting, or Github default branch, so that,

If branch is not set in .updatebot.yml, tries to resolve default branch name from remote Github repository. Otherwise, use 'master' as fallback for Git repositories, so that,

downstream repositories are cloned and checked out using remote branch name, so that,

pull requests are created to merge into the same remote branch.